Fixed some issues with castings on the C API
--HG--
branch : ode

+++ include/ode/common.h
@@ -122,15 +122,15 @@ typedef dReal dQuaternion[4];
 #if defined(dSINGLE)
 
 #define REAL(x) (x ## f)					/* form a constant */
-#define dRecip(x) ((float)(1.0f/(x)))				/* reciprocal */
-#define dSqrt(x) ((float)sqrtf(float(x)))			/* square root */
-#define dRecipSqrt(x) ((float)(1.0f/sqrtf(float(x))))		/* reciprocal square root */
-#define dSin(x) ((float)sinf(float(x)))				/* sine */
-#define dCos(x) ((float)cosf(float(x)))				/* cosine */
-#define dFabs(x) ((float)fabsf(float(x)))			/* absolute value */
-#define dAtan2(y,x) ((float)atan2f(float(y),float(x)))		/* arc tangent with 2 args */
-#define dFMod(a,b) ((float)fmodf(float(a),float(b)))		/* modulo */
-#define dCopySign(a,b) ((float)copysignf(float(a),float(b)))
+#define dRecip(x) ((1.0f/(x)))				/* reciprocal */
+#define dSqrt(x) (sqrtf(x))			/* square root */
+#define dRecipSqrt(x) ((1.0f/sqrtf(x)))		/* reciprocal square root */
+#define dSin(x) (sinf(x))				/* sine */
+#define dCos(x) (cosf(x))				/* cosine */
+#define dFabs(x) (fabsf(x))			/* absolute value */
+#define dAtan2(y,x) (atan2f(y,x))		/* arc tangent with 2 args */
+#define dFMod(a,b) (fmodf(a,b))		/* modulo */
+#define dCopySign(a,b) (copysignf(a,b))
 
 #elif defined(dDOUBLE)
